+++ src/scripts/constants.py
@@ -27,6 +27,7 @@ MAX_REPETITIONS = 50
 
 CHOICES_MAPPING = {
     "bg": "Възможности",
+    "ca": "Opcions",
     "cs": "Výběr",
     "da": "Svarmuligheder",
     "de": "Antwortmöglichkeiten",