+++ src/scripts/create_mmlu.py
@@ -31,6 +31,7 @@ from sklearn.model_selection import train_test_split
 LANGUAGES = [
     "da",
     "de",
+    "ca",
     "en",
     "es",
     "fr",