+++ src/scripts/create_multi_wiki_qa.py
@@ -52,6 +52,7 @@ LANGUAGES = [
     "hr",
     "bs",
     "hu",
+    "ca",
 ]