+++ src/scripts/create_scala.py
@@ -23,6 +23,7 @@ from datasets.dataset_dict import DatasetDict
 from huggingface_hub.hf_api import HfApi
 from load_ud_pos import (
     load_bgdt_pos,
+    load_cadt_pos,
     load_csdt_pos,
     load_dadt_pos,
     load_dedt_pos,
@@ -85,6 +86,7 @@ def main() -> None:
         "sl": load_sldt_pos,
         "hr": load_hrdt_pos,
         "hu": load_hudt_pos,
+        "ca": load_cadt_pos,
     }
 
     # Set up the progress bar and iterate over the languages